Fix circular dependency in VesselList component
