+++ frontend/src/features/Vessel/components/VesselList/VesselListModal.tsx
@@ -1,9 +1,5 @@
 import { StyledModalHeader } from '@features/commonComponents/StyledModalHeader'
 import { PrimaryButton, SecondaryButton } from '@features/commonStyles/Buttons.style'
-import { DownloadVesselListModal } from '@features/Vessel/components/VesselList/DownloadVesselListModal'
-import { VesselIcon } from '@features/Vessel/components/VesselList/index'
-import { VesselListFilters } from '@features/Vessel/components/VesselList/VesselListFilters'
-import { VesselListTable } from '@features/Vessel/components/VesselList/VesselListTable'
 import { vesselSelectors } from '@features/Vessel/slice'
 import { previewVessels } from '@features/Vessel/useCases/previewVessels'
 import { SaveVesselFiltersModal } from '@features/VesselFilter/components/SaveVesselFiltersModal'
@@ -14,6 +10,10 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { Modal } from 'rsuite'
 import styled from 'styled-components'
 
+import { DownloadVesselListModal } from './DownloadVesselListModal'
+import { VesselIcon } from './shared'
+import { VesselListFilters } from './VesselListFilters'
+import { VesselListTable } from './VesselListTable'
 import { setBlockVesselsUpdate } from '../../../../domain/shared_slices/Global'
 import { getFilteredVessels } from '../../../../domain/use_cases/vessel/getFilteredVessels'
 import { unselectVessel } from '../../../../domain/use_cases/vessel/unselectVessel'
