+++ frontend/src/features/Vessel/components/VesselList/index.tsx
@@ -1,7 +1,5 @@
 import { resetInteraction } from '@features/Draw/slice'
 import { MapToolButton } from '@features/MainWindow/components/MapButtons/shared/MapToolButton'
-import { addZoneSelected, reset } from '@features/Vessel/components/VesselList/slice'
-import { VesselListModal } from '@features/Vessel/components/VesselList/VesselListModal'
 import { useListenForDrawedGeometry } from '@hooks/useListenForDrawing'
 import { useMainAppDispatch } from '@hooks/useMainAppDispatch'
 import { useMainAppSelector } from '@hooks/useMainAppSelector'
@@ -10,12 +8,14 @@ import { useCallback, useEffect } from 'react'
 import { Modal } from 'rsuite'
 import styled from 'styled-components'
 
+import { VesselIcon } from './shared'
+import { addZoneSelected, reset } from './slice'
+import { VesselListModal } from './VesselListModal'
 import { LayerType } from '../../../../domain/entities/layers/constants'
 import { InteractionListener } from '../../../../domain/entities/map/constants'
 import { setDisplayedComponents } from '../../../../domain/shared_slices/DisplayedComponent'
 import { setBlockVesselsUpdate } from '../../../../domain/shared_slices/Global'
 import { MapComponent } from '../../../commonStyles/MapComponent'
-import VesselListSVG from '../../../icons/Icone_liste_navires.svg?react'
 
 export function VesselList({ namespace }) {
   const dispatch = useMainAppDispatch()
@@ -89,19 +89,3 @@ const Wrapper = styled(MapComponent)`
 `
 
 const VesselListButton = styled(MapToolButton)``
-
-export const VesselIcon = styled(VesselListSVG)<{
-  $background: string
-  $isRightMenuShrinked: boolean | undefined
-  $isTitle: boolean
-}>`
-  width: 25px;
-  height: 25px;
-  margin-top: 4px;
-  opacity: ${p => (p.$isRightMenuShrinked ? '0' : '1')};
-  vertical-align: ${p => (p.$isTitle ? 'text-bottom' : 'baseline')};
-  circle {
-    fill: ${p => p.$background};
-  }
-  transition: all 0.3s;
-`
